requirements.txt
pandas
kagglehub
python-dotenv
psycopg[binary]
Flask
